+++ Core/OneLog.Domain/Entities/Log.cs
@@ -5,10 +5,10 @@ namespace OneLog.Domain.Entities
 {
     public class Log : BaseEntity
     {
-        public int ApplicationId { get; set; }
+        public long ApplicationId { get; set; }
         [ForeignKey("ApplicationId")]
         public BusinessApplication? Application { get; set; }
-        public int BusinessId { get; set; }
+        public long BusinessId { get; set; }
         [ForeignKey("BusinessId")]
         public Business? Business { get; set; }
         public string Message { get; set; } = string.Empty;
